requirements.txt
asgiref==3.2.7
aioredis==1.3.1
Django==3.0.6
async-timeout==3.0.1
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
python-crontab==2.5.1
celery==4.4.6
channels==2.4.0
channels-redis==2.4.2
daphne==2.5.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
eventlet==0.25.2
redis==3.5.3
websocket-client==0.57.0
websockets==8.1
gunicorn==20.0.4
Pillow==7.1.2
whitenoise==5.0.1